+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Features
 
+- Added a unified experiment-model path that reuses a single built model and solver across compatible experiment steps. ([#5422](https://github.com/pybamm-team/PyBaMM/pull/5422))
+
 ## Bug fixes
 
 ## Breaking changes